+++ src/Author.java
@@ -1,0 +1,9 @@
+public class Author{
+    private String authorName;
+    public Author(String authorName) {
+        this.authorName = authorName;
+    }
+    public String getAuthorName() {
+        return this.authorName;
+    }
+}
